move divider into staticfilters
