+++ sample-app/src/components/StaticFilters.tsx
@@ -1,7 +1,6 @@
 import { useAnswersActions, useAnswersState, Filter, Matcher } from '@yext/answers-headless-react';
 import { CompositionMethod, useComposedCssClasses } from '../hooks/useComposedCssClasses';
 import { isDuplicateFilter } from '../utils/filterutils';
-import Divider from './Divider';
 
 interface CheckBoxProps {
   fieldId: string,
@@ -117,4 +116,19 @@ function CheckboxFilter({ fieldId, value, label, selected, optionHandler, cssCla
       <label className={cssClasses.optionLabel} htmlFor={id}>{label}</label>
     </div>
   );
+}
+
+interface DividerProps {
+  customCssClasses?: {
+    divider?: string
+  },
+  cssCompositionMethod?: CompositionMethod
+}
+
+export function Divider({ customCssClasses, cssCompositionMethod }: DividerProps) {
+  const builtInCssClasses = {
+    divider: 'w-full h-px bg-gray-200 my-4'
+  }
+  const cssClasses = useComposedCssClasses(builtInCssClasses, customCssClasses, cssCompositionMethod);
+  return <div className={cssClasses.divider}></div>
 }
